robot_agent_prototype_leonardo/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
pydantic>=2.9,<3.0
pyserial>=3.5,<4.0
